scripts/requirements.txt
zulip == 0.8.2
PyGithub==1.55
python-decouple==3.6
rich==12.2.0
littletable== 2.0.6
thefuzz == 0.19.0
loguru == 0.6.0 # notifiers integrated
python-Levenshtein==0.12.2
twilio==7.10.0